# perf: split sos_square loop